trend-score/requirements.txt
streamlit
pandas
numpy
alpaca-trade-api
ta
joblib
